+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This changelog consists of the bug & security fixes and new features being inclu
 
 ## Unreleased
 
+* Added a method in the Installer class to prompt for a GitHub star once the installation is completed.
+
 * #10782 [fixed] - Resolved an issue where the "Bundle Items" section was not visible while creating a Bundle Product in the French locale (APP_LOCALE=fr). This was caused by unescaped apostrophes in translatable strings breaking JavaScript during rendering.
 
 ## **v2.2.9 (11th of June 2025)** - *Release*